backend/requirements/base.txt
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
redis==5.0.6
djangorestframework-simplejwt==5.3.1
pillow==10.3.0